refactor: EnumType 인터페이스를 common 패키지 하위로 이동한다.
