+++ src/main/java/com/clova/anifriends/domain/applicant/wrapper/ApplicantStatus.java
@@ -1,6 +1,6 @@
 package com.clova.anifriends.domain.applicant.wrapper;
 
-import com.clova.anifriends.EnumType;
+import com.clova.anifriends.domain.common.EnumType;
 
 public enum ApplicantStatus implements EnumType {
